+++ src/app/look-at-gaming/page.tsx
@@ -5,12 +5,12 @@ import React, {
   useState
 } from "react";
 import TCL_Logo from "@/components/core/TCL_Logo";
-import LAG_Create from "@/components/create_lag/LAG_Create";
-import LAG_Preface from "@/components/create_lag/LAG_Preface";
-import LAG_Preview from "@/components/create_lag/LAG_Preview";
+import LAG_Create from "@/components/look-at-gaming/LAG_Create";
+import LAG_Preface from "@/components/look-at-gaming/LAG_Preface";
+import LAG_Preview from "@/components/look-at-gaming/LAG_Preview";
 import { ILAG } from "@/utils/types";
 import { LAG_CATEGORIES } from "@/utils/lag-categories";
-import LAG_SAMPLE from "@/utils/lag-samples";
+// import LAG_SAMPLE from "@/utils/lag-samples";
 
 // LAG Context
 const LookAtGaming = createContext({});
@@ -29,15 +29,14 @@ export default function Create_LAG(): React.ReactElement {
     footer: "",
   };
 
-  // const [LAG, setLAG] = useState<ILAG>(LAG_EMPTY);
-  const [LAG, setLAG] = useState<ILAG>(LAG_SAMPLE);
+  const [LAG, setLAG] = useState<ILAG>(LAG_EMPTY);
 
   useEffect(() => {
     console.log("LAG: ", LAG);
   }, [LAG]);
 
   return (
-    <LookAtGaming.Provider value={{LAG, setLAG}}>
+    <LookAtGaming.Provider value={{ LAG, setLAG }}>
       <main className="root-container">
         <TCL_Logo />
