+++ src/components/look-at-gaming/LAG_Create.module.css
@@ -10,7 +10,7 @@
   .row {
     display: flex;
     flex-direction: row;
-    justify-content: start;
+    justify-content: flex-start;
     align-items: center;
     gap: 1rem;
     min-height: 40px;
@@ -20,7 +20,7 @@
 .label {
   display: flex;
   flex-direction: row;
-  justify-content: start;
+  justify-content: flex-start;
   align-items: center;
   margin: 0;
   padding: 0;
@@ -46,8 +46,8 @@
 .subheading-container {
   display: flex;
   flex-direction: column;
-  justify-content: start;
-  align-items: start;
+  justify-content: flex-start;
+  align-items: flex-start;
   width: 100%;
 
   .textarea {
@@ -63,8 +63,8 @@
 .footer-container {
   display: flex;
   flex-direction: column;
-  justify-content: start;
-  align-items: start;
+  justify-content: flex-start;
+  align-items: flex-start;
   width: 100%;
 
   .textarea {
@@ -80,16 +80,16 @@
 .category-group-container {
   display: flex;
   flex-direction: column;
-  justify-content: start;
-  align-items: start;
+  justify-content: flex-start;
+  align-items: flex-start;
   gap: 3rem;
   width: 100%;
 
   .category-container {
     display: flex;
     flex-direction: column;
-    justify-content: start;
-    align-items: start;
+    justify-content: flex-start;
+    align-items: flex-start;
     gap: 0.5rem;
     position: relative;
     width: 100%;
@@ -101,7 +101,7 @@
     .category {
       display: flex;
       flex-direction: row;
-      justify-content: start;
+      justify-content: flex-start;
       align-items: center;
       gap: 1rem;
       width: 100%;
@@ -121,8 +121,8 @@
     .article-container {
       display: flex;
       flex-direction: column;
-      justify-content: start;
-      align-items: start;
+      justify-content: flex-start;
+      align-items: flex-start;
       gap: 1rem;
       width: 100%;
       box-sizing: border-box;
@@ -130,8 +130,8 @@
       .article {
         display: flex;
         flex-direction: row;
-        justify-content: start;
-        align-items: start;
+        justify-content: flex-start;
+        align-items: flex-start;
         width: 100%;
         border: none;
         border-radius: 4px;
@@ -143,8 +143,9 @@
           .row {
             display: flex;
             flex-direction: row;
-            justify-content: start;
+            justify-content: flex-start;
             align-items: center;
+            gap: 0;
             width: 100%;
 
             .icon {
@@ -212,28 +213,6 @@
 
     }
   }
-}
-
-.create-section button {
-  padding: 0.5rem 1rem;
-  color: white;
-  font-size: 16px;
-  font-weight: bold;
-  border: 1px solid transparent;
-  border-radius: 4px;
-  cursor: pointer;
-  box-sizing: border-box;
-  transition: all 100ms linear;
-}
-
-.create-section button:focus,
-.create-section button:hover {
-  outline: none;
-  filter: brightness(80%);
-}
-.create-section button:active {
-  box-shadow: none;
-  filter: brightness(70%);
 }
 
 .create-section button.add {
